Simplify query engine under assumption of type inference (#30)
We simplify and remove unneeded code branches that were created to handle the case where Type Inference has not run. We build in the assumption that all Thing vertices will have a set of possible types added via type inference (but not for Type vertices, which may not for schema-only queries). In the process we also move the responsibility of resolving role instance types from Planner/Procedure to Type Inference, where it should belong.

* Delete Planner code paths that defended against nonexistent inferred types for Thing vertices, we now always assume there are a set of possible types populated
* Delete Procedure code paths that defended against nonexistent inferred types for Thing vertices, we now always assume there are a set of possible types populated
* Extract duplicated type inference for role instances from the Planner and Procedure to Type Inference, and pass the result into the traversal engine
* Add a new rule validation for ambiguous type labels (ie. roles) that should not be allowed in a rule conclusion. Consider `then { (role: $x) isa marriage;}` where the `role` could be either `husband` or `wife` - it is ambiguous
* Fix various tests following from these changes
* Rename `Conjunction.Cloner` to `Conjunction.ConstraintCloner` since it is used to clone a subset of conjunction constraints from the conjunction

put ORDER parameter last in all Sorted/Seekable iterator methods

Add notes for optimisations with seek()